+++ CHANGELOG.md
@@ -13,6 +13,7 @@ the APIs in these libraries are stable, and are ready for production use.
 
 - [Memorystore API](/google/cloud/memorystore/README.md)
 - [Oracle Database@Google Cloud API](/google/cloud/oracledatabase/README.md)
+- [Dataform API](/google/cloud/dataform/README.md)
 
 ## v2.37.0 - 2025-05